requirements.txt
alembic==1.14.0
bcrypt==4.2.1
beautifulsoup4==4.12.3
blinker==1.9.0
certifi==2024.12.14
charset-normalizer==3.4.0
click==8.1.7
dnspython==2.7.0
email_validator==2.2.0
Flask==3.1.0
Flask-Bcrypt==1.0.1
Flask-Dance==7.1.0
Flask-Login==0.6.3
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
frozendict==2.4.6
gunicorn==23.0.0
html5lib==1.1
idna==3.10
importlib_metadata==8.5.0
itsdangerous==2.2.0
Jinja2==3.1.4
lxml==5.3.0
Mako==1.3.8
MarkupSafe==3.0.2
multitasking==0.0.11
numpy==2.0.2
oauthlib==3.2.2
packaging==24.2
pandas==2.2.3
peewee==3.17.8
platformdirs==4.3.6
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
requests==2.32.3
requests-oauthlib==2.0.0
six==1.17.0
soupsieve==2.6
SQLAlchemy==2.0.36
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
URLObject==2.4.3
webencodings==0.5.1
Werkzeug==3.1.3
WTForms==3.2.1
yfinance==0.2.50
zipp==3.21.0
